backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.7            # via django
django-cors-headers==3.3.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django==3.0.7             # via -r requirements.in, django-cors-headers, django-filter, djangorestframework, djangorestframework-simplejwt
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.11.0  # via -r requirements.in, djangorestframework-simplejwt
pyjwt==1.7.1              # via djangorestframework-simplejwt
pytz==2020.1              # via django
sqlparse==0.3.1           # via django
